requirements.txt
chromadb
imagehash
numpy
openai
paddleocr
pillow
psutil
pydbus
PyYAML
voyageai
pyarrow
scipy
# Agents dependencies
pytz
pandas
openpyxl
Markdown
scikit-image
tiktoken
google-genai
python-dotenv
demjson3
pathvalidate
docstring_parser
sqlalchemy
pydantic-settings
jinja2
humps
composio
colorama
anthropic
httpx_sse
rapidfuzz
rank-bm25
llama_index
llama-index-embeddings-google-genai
fastapi>=0.104.1
uvicorn[standard]>=0.31.1
pydub
python-multipart
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests
SpeechRecognition
ffmpeg
pg8000
asyncpg
pgvector
json_repair
rich>=13.7.1,<14.0.0
psycopg2-binary
anyio>=4.7.0
mcp
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
ruff
nodeenv
